move moduleconfig -> service/moduleconfig
