+++ internal/service/moduleconfig/moduleconfig.go
@@ -17,12 +17,12 @@ type FileGenerator interface {
 	GenerateFile(out io.Out, path string, args types.KeyValueArgs) error
 }
 
-type ModuleConfigService struct {
+type Service struct {
 	fileSystem    FileSystem
 	fileGenerator FileGenerator
 }
 
-func NewModuleConfigService(fileSystem FileSystem, fileGenerator FileGenerator) (*ModuleConfigService, error) {
+func NewService(fileSystem FileSystem, fileGenerator FileGenerator) (*Service, error) {
 	if fileSystem == nil {
 		return nil, fmt.Errorf("%w: fileSystem must not be nil", errors.ErrInvalidArg)
 
@@ -32,13 +32,13 @@ func NewModuleConfigService(fileSystem FileSystem, fileGenerator FileGenerator) 
 		return nil, fmt.Errorf("%w: fileGenerator must not be nil", errors.ErrInvalidArg)
 	}
 
-	return &ModuleConfigService{
+	return &Service{
 		fileSystem:    fileSystem,
 		fileGenerator: fileGenerator,
 	}, nil
 }
 
-func (s *ModuleConfigService) PreventOverwrite(directory, fileName string, overwrite bool) error {
+func (s *Service) PreventOverwrite(directory, fileName string, overwrite bool) error {
 	exists, err := s.fileSystem.FileExists(path.Join(directory, fileName))
 	if err != nil {
 		return err
@@ -51,6 +51,6 @@ func (s *ModuleConfigService) PreventOverwrite(directory, fileName string, overw
 	return nil
 }
 
-func (s *ModuleConfigService) GenerateFile(out io.Out, path string, args types.KeyValueArgs) error {
+func (s *Service) GenerateFile(out io.Out, path string, args types.KeyValueArgs) error {
 	return s.fileGenerator.GenerateFile(out, path, args)
 }
