+++ internal/service/moduleconfig/moduleconfig_test.go
@@ -9,7 +9,7 @@ import (
 
 	commonerrors "github.com/kyma-project/modulectl/internal/scaffold/common/errors"
 	"github.com/kyma-project/modulectl/internal/scaffold/common/types"
-	"github.com/kyma-project/modulectl/internal/scaffold/moduleconfig"
+	"github.com/kyma-project/modulectl/internal/service/moduleconfig"
 	"github.com/kyma-project/modulectl/tools/io"
 )
 
@@ -18,8 +18,8 @@ const (
 	moduleConfigFile = "config.yaml"
 )
 
-func Test_NewModuleConfigService_ReturnsError_WhenFileSystemIsNil(t *testing.T) {
-	_, err := moduleconfig.NewModuleConfigService(
+func Test_NewService_ReturnsError_WhenFileSystemIsNil(t *testing.T) {
+	_, err := moduleconfig.NewService(
 		nil,
 		&fileGeneratorErrorStub{},
 	)
@@ -28,8 +28,8 @@ func Test_NewModuleConfigService_ReturnsError_WhenFileSystemIsNil(t *testing.T) 
 	assert.Contains(t, err.Error(), "fileSystem must not be nil")
 }
 
-func Test_NewModuleConfigService_ReturnsError_WhenFileGeneratorIsNil(t *testing.T) {
-	_, err := moduleconfig.NewModuleConfigService(
+func Test_NewService_ReturnsError_WhenFileGeneratorIsNil(t *testing.T) {
+	_, err := moduleconfig.NewService(
 		&errorStub{},
 		nil,
 	)
@@ -39,7 +39,7 @@ func Test_NewModuleConfigService_ReturnsError_WhenFileGeneratorIsNil(t *testing.
 }
 
 func Test_PreventOverwrite_ReturnsError_WhenFilesystemReturnsError(t *testing.T) {
-	svc, _ := moduleconfig.NewModuleConfigService(
+	svc, _ := moduleconfig.NewService(
 		&errorStub{},
 		&fileGeneratorErrorStub{},
 	)
@@ -50,7 +50,7 @@ func Test_PreventOverwrite_ReturnsError_WhenFilesystemReturnsError(t *testing.T)
 }
 
 func Test_PreventOverwrite_ReturnsError_WhenFileExistsAndNoOverwrite(t *testing.T) {
-	svc, _ := moduleconfig.NewModuleConfigService(
+	svc, _ := moduleconfig.NewService(
 		&fileExistsStub{},
 		&fileGeneratorErrorStub{},
 	)
@@ -61,7 +61,7 @@ func Test_PreventOverwrite_ReturnsError_WhenFileExistsAndNoOverwrite(t *testing.
 }
 
 func Test_PreventOverwrite_ReturnsNil_WhenFileExistsAndOverwrite(t *testing.T) {
-	svc, _ := moduleconfig.NewModuleConfigService(
+	svc, _ := moduleconfig.NewService(
 		&fileExistsStub{},
 		&fileGeneratorErrorStub{},
 	)
@@ -72,7 +72,7 @@ func Test_PreventOverwrite_ReturnsNil_WhenFileExistsAndOverwrite(t *testing.T) {
 }
 
 func Test_PreventOverwrite_ReturnsNil_WhenFileDoesNotExist(t *testing.T) {
-	svc, _ := moduleconfig.NewModuleConfigService(
+	svc, _ := moduleconfig.NewService(
 		&fileDoesNotExistStub{},
 		&fileGeneratorErrorStub{},
 	)
@@ -83,7 +83,7 @@ func Test_PreventOverwrite_ReturnsNil_WhenFileDoesNotExist(t *testing.T) {
 }
 
 func Test_GenerateFile_ReturnsError_WhenFileGeneratorReturnsError(t *testing.T) {
-	svc, _ := moduleconfig.NewModuleConfigService(
+	svc, _ := moduleconfig.NewService(
 		&errorStub{},
 		&fileGeneratorErrorStub{},
 	)
@@ -94,7 +94,7 @@ func Test_GenerateFile_ReturnsError_WhenFileGeneratorReturnsError(t *testing.T) 
 }
 
 func Test_GenerateFile_Succeeds(t *testing.T) {
-	svc, _ := moduleconfig.NewModuleConfigService(
+	svc, _ := moduleconfig.NewService(
 		&errorStub{},
 		&fileGeneratorStub{},
 	)
